+++ src/html/index.html
@@ -9,9 +9,10 @@
 <body>
   <div id="content">
     <h1>hola</h1>
-    <span id="something"></span>
+    <span id="something"></span><br>
+    <span id="somethingElse"></span>
   </div>
 
-  <!-- <script src="../js/preload.js"></script> -->
+  <script src="../js/index.js"></script>
 </body>
 </html>
