--- src/js/defaults.js
@@ -1,10 +1,0 @@
-
-module.exports = [
-  {
-    name: 'WhatsApp',
-    logo: '../images/logo.png',
-    url: 'https://web.whatsapp.com',
-    width: 800,
-    height: 600
-  }
-]
