+++ src/js/index.js
@@ -1,0 +1,4 @@
+console.log('hello from index')
+
+otherThingSpan = document.getElementById('somethingElse')
+otherThingSpan.innerHTML = 'other thing span from index'
